requirements.txt
flask>=3.0
flask-migrate>=4.0
flask-sqlalchemy>=3.1
sqlalchemy>=2.0
weasyprint>=60.0
python-dotenv>=1.0
